+++ Assets/SchoolRPG/Source/Framework/Player/health/health_controller.cs
@@ -1,0 +1,22 @@
+using System.Collections;
+using System.Collections.Generic;
+using static health.health;
+using static death.Death; 
+using UnityEngine;
+namespace healthController
+{
+    public class health_controller : MonoBehaviour
+    {
+        // Start is called before the first frame update
+        void Start()
+        {
+
+        }
+
+        // Update is called once per frame
+        void Update()
+        {
+
+        }
+    }
+}
