+++ Assets/SchoolRPG/Source/Framework/Player/health/health_controller.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 38f430e967b9184468c6f30505beba2b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
